feat(widgets): 提取 UnifiedReferencePanel 组件，减少文件行数
将 unified_reference_panel.dart 从 1052 行缩减至 264 行（目标 <500行）

提取的组件：
- drag_target_wrapper.dart: DragTarget 包装器，处理拖拽逻辑和库操作
- recent_vibes_list.dart: 最近使用的 Vibes 列表组件
- library_actions_row.dart: 库操作按钮行（保存到库、从库导入）

优化：
- 使用静态方法供外部调用添加 Vibe 功能
- 保持原有功能完整性
- 代码结构更清晰，便于维护
